+++ CHANGELOG.md
@@ -65,6 +65,7 @@
 * Add Hub search behavior for LoC source; fix search results table. Refs [UILD-688].
 * Add Hubs actions menu. Refs [UILD-693].
 * Add initial workspace profile settings editing. Refs [UILD-685].
+* Refactor Edit page to support multiple resource types. Refs [UILD-694]. 
 
 [UILD-552]:https://folio-org.atlassian.net/browse/UILD-552
 [UILD-544]:https://folio-org.atlassian.net/browse/UILD-544
@@ -130,6 +131,7 @@
 [UILD-688]:https://folio-org.atlassian.net/browse/UILD-688
 [UILD-693]:https://folio-org.atlassian.net/browse/UILD-693
 [UILD-685]:https://folio-org.atlassian.net/browse/UILD-685
+[UILD-694]:https://folio-org.atlassian.net/browse/UILD-694
 
 ## 1.0.5 (2025-04-30)
 * Fixed incorrect behavior when navigating between duplicated resources. Fixes [UILD-553].